feat: restructure packages to global scope and add logging components
