--- backend/src/main/kotlin/day/widdle/widdle/HealthController.kt
@@ -1,14 +1,0 @@
-package day.widdle.widdle
-
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/health")
-class HealthController {
-
-    @GetMapping
-    fun healthCheck(): ResponseEntity<String> = ResponseEntity.ok("OK")
-}
