+++ backend/src/main/kotlin/day/widdle/widdle/correction/service/checker/BareunSpellChecker.kt
@@ -5,8 +5,8 @@ import day.widdle.widdle.correction.service.KoreanSpellChecker
 import day.widdle.widdle.correction.service.dto.CorrectionResult
 import day.widdle.widdle.correction.service.dto.bareun.CorrectErrorRequest
 import day.widdle.widdle.correction.service.dto.bareun.CorrectErrorResponse
-import day.widdle.widdle.exception.WiddleException
-import day.widdle.widdle.support.logger
+import day.widdle.widdle.global.exception.WiddleException
+import day.widdle.widdle.global.support.logger
 import io.netty.handler.ssl.SslContextBuilder
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory
 import kotlinx.coroutines.reactor.awaitSingleOrNull
