+++ backend/src/main/kotlin/day/widdle/widdle/correction/service/checker/OpenAiChecker.kt
@@ -5,7 +5,7 @@ import day.widdle.widdle.correction.service.KoreanSpellChecker
 import day.widdle.widdle.correction.service.dto.CorrectionResult
 import day.widdle.widdle.correction.service.dto.openai.OpenAiRequest
 import day.widdle.widdle.correction.service.dto.openai.OpenAiResponse
-import day.widdle.widdle.exception.WiddleException
+import day.widdle.widdle.global.exception.WiddleException
 import kotlinx.coroutines.reactor.awaitSingleOrNull
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
