--- backend/src/main/kotlin/day/widdle/widdle/event/WiddleEvent.kt
@@ -1,3 +1,0 @@
-package day.widdle.widdle.event
-
-interface WiddleEvent
