--- backend/src/main/kotlin/day/widdle/widdle/event/WordSavedEvent.kt
@@ -1,3 +1,0 @@
-package day.widdle.widdle.event
-
-data class WordSavedEvent(val jamo: List<String>) : WiddleEvent
