--- backend/src/main/kotlin/day/widdle/widdle/event/listener/NewWordEventListener.kt
@@ -1,21 +1,0 @@
-package day.widdle.widdle.event.listener
-
-import day.widdle.widdle.event.NewWordEvent
-import day.widdle.widdle.support.logger
-import day.widdle.widdle.word.service.WordTransactionalService
-import org.springframework.context.event.EventListener
-import org.springframework.scheduling.annotation.Async
-import org.springframework.stereotype.Component
-
-@Component
-class NewWordEventListener(
-    private val wordTransactionalService: WordTransactionalService
-) {
-    private val log = logger()
-
-    @Async
-    @EventListener
-    fun processNewWordEvent(event: NewWordEvent) {
-        wordTransactionalService.save(event.word, event.jamo, event.isKorean)
-    }
-}
