--- backend/src/main/kotlin/day/widdle/widdle/exception/WiddleException.kt
@@ -1,3 +1,0 @@
-package day.widdle.widdle.exception
-
-class WiddleException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
