+++ backend/src/main/kotlin/day/widdle/widdle/global/HealthController.kt
@@ -1,0 +1,14 @@
+package day.widdle.widdle.global
+
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/health")
+class HealthController {
+
+    @GetMapping
+    fun healthCheck(): ResponseEntity<String> = ResponseEntity.ok("OK")
+}
