+++ backend/src/main/kotlin/day/widdle/widdle/global/annotation/LogExternal.kt
@@ -1,0 +1,5 @@
+package day.widdle.widdle.global.annotation
+
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class LogExternal()
