+++ backend/src/main/kotlin/day/widdle/widdle/global/event/WiddleEvent.kt
@@ -1,0 +1,3 @@
+package day.widdle.widdle.global.event
+
+interface WiddleEvent
