+++ backend/src/main/kotlin/day/widdle/widdle/global/event/WordSavedEvent.kt
@@ -1,0 +1,3 @@
+package day.widdle.widdle.global.event
+
+data class WordSavedEvent(val jamo: List<String>) : WiddleEvent
