+++ backend/src/main/kotlin/day/widdle/widdle/global/event/publisher/WiddleEventPublisher.kt
@@ -1,0 +1,22 @@
+package day.widdle.widdle.global.event.publisher
+
+import day.widdle.widdle.global.event.WiddleEvent
+import org.springframework.context.ApplicationEventPublisher
+import org.springframework.stereotype.Component
+
+@Component
+class WiddleEventPublisher(
+    private val publisher: ApplicationEventPublisher
+) {
+
+    fun publishEvent(flag: Boolean, widdleEvent: WiddleEvent) {
+        if (flag) {
+            publishEvent(widdleEvent)
+        }
+    }
+
+    fun publishEvent(widdleEvent: WiddleEvent) {
+        publisher.publishEvent(widdleEvent)
+    }
+
+}
