+++ backend/src/main/kotlin/day/widdle/widdle/global/exception/WiddleException.kt
@@ -1,0 +1,3 @@
+package day.widdle.widdle.global.exception
+
+class WiddleException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
