+++ backend/src/main/kotlin/day/widdle/widdle/global/support/LocalDateUtil.kt
@@ -1,0 +1,7 @@
+package day.widdle.widdle.global.support
+
+import java.time.LocalDate
+import java.time.ZoneId
+
+private val zone = ZoneId.of("Asia/Seoul")
+fun getToday(): LocalDate = LocalDate.now(zone)
