+++ backend/src/main/kotlin/day/widdle/widdle/global/support/Mapper.kt
@@ -1,0 +1,7 @@
+package day.widdle.widdle.global.support
+
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+
+fun <T> T.toResponse(status: HttpStatus = HttpStatus.OK): ResponseEntity<T> =
+    ResponseEntity.status(status).body(this)
