+++ backend/src/main/kotlin/day/widdle/widdle/global/support/StringUtils.kt
@@ -1,0 +1,7 @@
+package day.widdle.widdle.global.support
+
+import day.widdle.widdle.global.support.JamoSplitter.splitToJamoOrChar
+
+fun String.isKorean(): Boolean = this.toCharArray().all { char -> char.code in 0xAC00..0xD7A3 }
+
+fun String.toJamoList(): List<String> = splitToJamoOrChar(this, this.isKorean())
