+++ backend/src/main/kotlin/day/widdle/widdle/search/service/SearchService.kt
@@ -1,8 +1,8 @@
 package day.widdle.widdle.search.service
 
-import day.widdle.widdle.event.NewWordEvent
-import day.widdle.widdle.event.publisher.WiddleEventPublisher
-import day.widdle.widdle.support.logger
+import day.widdle.widdle.global.event.NewWordEvent
+import day.widdle.widdle.global.event.publisher.WiddleEventPublisher
+import day.widdle.widdle.global.support.logger
 import day.widdle.widdle.search.value.DictionaryType
 import day.widdle.widdle.search.value.getDictionaryType
 import org.springframework.stereotype.Service
