+++ backend/src/main/kotlin/day/widdle/widdle/search/value/DictionaryType.kt
@@ -2,7 +2,7 @@ package day.widdle.widdle.search.value
 
 import day.widdle.widdle.search.value.DictionaryType.ENGLISH
 import day.widdle.widdle.search.value.DictionaryType.KOREAN
-import day.widdle.widdle.support.isKorean
+import day.widdle.widdle.global.support.isKorean
 
 enum class DictionaryType {
     NAVER, ENGLISH, KOREAN;
