--- backend/src/main/kotlin/day/widdle/widdle/support/LocalDateUtil.kt
@@ -1,7 +1,0 @@
-package day.widdle.widdle.support
-
-import java.time.LocalDate
-import java.time.ZoneId
-
-private val zone = ZoneId.of("Asia/Seoul")
-fun getToday(): LocalDate = LocalDate.now(zone)
