--- backend/src/main/kotlin/day/widdle/widdle/support/Mapper.kt
@@ -1,7 +1,0 @@
-package day.widdle.widdle.support
-
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-
-fun <T> T.toResponse(status: HttpStatus = HttpStatus.OK): ResponseEntity<T> =
-    ResponseEntity.status(status).body(this)
