--- backend/src/main/kotlin/day/widdle/widdle/support/StringUtils.kt
@@ -1,7 +1,0 @@
-package day.widdle.widdle.support
-
-import day.widdle.widdle.support.JamoSplitter.splitToJamoOrChar
-
-fun String.isKorean(): Boolean = this.toCharArray().all { char -> char.code in 0xAC00..0xD7A3 }
-
-fun String.toJamoList(): List<String> = splitToJamoOrChar(this, this.isKorean())
